+++ src/array/codec/array_to_bytes/sharding/sharding_partial_decoder.rs
@@ -1,6 +1,7 @@
 use std::{num::NonZeroU64, sync::Arc};
 
 use rayon::prelude::*;
+use unsafe_cell_slice::UnsafeCellSlice;
 
 use crate::{
     array::{
@@ -13,9 +14,8 @@ use crate::{
             CodecOptions,
         },
         concurrency::{calc_concurrency_outer_inner, RecommendedConcurrency},
-        ravel_indices,
-        unsafe_cell_slice::UnsafeCellSlice,
-        ArrayBytes, ArraySize, ChunkRepresentation, ChunkShape, DataType, DataTypeSize,
+        ravel_indices, ArrayBytes, ArraySize, ChunkRepresentation, ChunkShape, DataType,
+        DataTypeSize,
     },
     byte_range::ByteRange,
 };
@@ -324,7 +324,7 @@ impl ArrayPartialDecoderTraits for ShardingPartialDecoder<'_> {
                         };
                         let decoded_bytes = decoded_bytes.into_fixed()?;
                         update_bytes_flen(
-                            unsafe { out_array_subset_slice.get() },
+                            unsafe { out_array_subset_slice.as_mut_slice() },
                             array_subset.shape(),
                             &decoded_bytes,
                             &chunk_subset_overlap
@@ -668,7 +668,7 @@ impl AsyncArrayPartialDecoderTraits for AsyncShardingPartialDecoder<'_> {
                                     ArraySubset,
                                 ) = subset_and_decoded_chunk?;
                                 update_bytes_flen(
-                                    unsafe { shard_slice.get() },
+                                    unsafe { shard_slice.as_mut_slice() },
                                     array_subset.shape(),
                                     &chunk_subset_bytes.into(),
                                     &chunk_subset_overlap
@@ -707,7 +707,7 @@ impl AsyncArrayPartialDecoderTraits for AsyncShardingPartialDecoder<'_> {
                                     .as_ne_bytes()
                                     .repeat(chunk_subset_overlap.num_elements_usize());
                                 update_bytes_flen(
-                                    unsafe { shard_slice.get() },
+                                    unsafe { shard_slice.as_mut_slice() },
                                     array_subset.shape(),
                                     &filled_chunk.into(),
                                     &chunk_subset_overlap